renamed files and var names for routing and finding out easy direct paths to be called to login the admin
